Graph updates now automatically on regex change
After changing a "Hostname Regex", the graph was not automatically
updated, but Grafanas "Refresh dashboard" button on the upper right had
to be clicked. Now the graph should update automatically. In order to
limit the load on the checkmk server, the update is delayed by 500ms for
text fields.

This affects:
* Hostname regex
* Service Regex
* Host is in Group
* Service is in Group