Schedule CI to run every day
Current CI failures let me remember that we wanted to use the latest
docker images in conjunction with a daily run of our CI to detect
upstream changes and compatibility issues.

CI is now schedule at 38 minutes after UTC midnight. This is a random
number of minutes to lower the load at full hours.