'Retry' time-usage panel test
We hereby modify the time-usage panel test by introducing the 'retries'
parameter. In this way, we deal with the tests flakyness in a more
general way by repeating it up to 3 times in case of failure.

Considering this, we avoid the usage of the 'recursiveType' function,
thus making the test more readable.