Generate random data in rrd files
The checkmk container runs only for seconds before the e2e tests run, so
no data is available when displaying metrics from this container.

The script added watches for rrds beeing created and generates some
random data, so metrics for the last 5 hours are available, even the
container was started just seconds before.