fix rrdtool create start argument
documentation for the `--start` argument says:
   Specifies the time in seconds since 1970-01-01 UTC when the first
   value should be added to the RRD

But this seems to be wrong. A 0 is interpreted as 00:00 because the
input is tried to be parsed as "at-style". As we go back 5 hours,
schedule the daily test-build around midnight, and rrd can not write
data before `--start` the update fails.
The solution is to specify some random date in the past.
I've choosen 2022-01-01