remove usage of generic default value
the original intent was to to make sure that Value is bound to
RequestSpec[Key], but actually typescript was thinking of Value as
completely untyped, although they way we use it currently it's alway
`string` or `string | undefined`.
by removing this default completely we don't longer need to cast value
to string.