move error message rewrite to source of errors
before "explore" and the graph panel did show the original (non
rewritten) error message.

currently we have two problems with that implementation:
1. the error message is translated, so this only works in english
   instances of checkmk
2. we duplicated the code between rest and web api.

for 1. we plan to introduce a feature in the REST-API (but this will
only solve the problem in `src/backend/rest.ts` not for the webapi!), we
probably have to live with the code duplication.