fix daily build: failing grafana plugin validator
Grafana plugin validator compares the current version with the latest
published version and will only pass if the current version is greater
than the latest published version.

Before this change we only added `-testbuild.XXX` to the version of the
plugin. And used info.version in `src/plugin.json`, but this only
holds a placeholder. So if we want to increase the version we need to
load version (not the placeholder) which is stored in package.json.