Clarify error message on 404 on REST-API request
We assumed that this can only happen if you query a Checkmk version that
has no graph query endpoints at all. But the enterprise REST-API
endpoint will return a 404 error in Checkmk raw edition, so we adapted
the hint for the user.

Unrelated fix: Checkmk should start with capital letter.